Fix: zone distribution percentages don't sum to 100% due to integer rounding
calculateZoneDistribution was computing (count/total*100), rounding to an
integer with Math.round(), then dividing back by 100f. For 3 equal zones of
33.33% each: round(33.33)=33, 33/100f=0.33, sum=0.99 instead of 1.0.

This rounding error accumulates in DashboardFragment where zoneDistribution
values are summed across multiple sessions.

Store count/total directly as a clean 0.0-1.0 fraction with full float
precision. Any display rounding belongs at the UI layer, not in storage.

+++ app/src/main/java/com/rouf/saht/heartRate/view/HeartRateAdapter.kt
@@ -64,7 +64,7 @@ class HeartRateAdapter(private val context: Context) : RecyclerView.Adapter<Hear
         init {
             itemView.setOnClickListener {
                 val pos = bindingAdapterPosition
-                if (pos != RecyclerView.NO_ID) onItemClick(pos)
+                if (pos.toLong() != RecyclerView.NO_ID) onItemClick(pos)
             }
         }
